Move script utils
